fix: resolve all real Swift compile errors across Core/UI, Features, Sources
Swift language mode set to 5.0 in project.yml (codebase is pre-Swift-6;
strict concurrency migration is a planned task, not a fire).

Core/UI (9): isLoading redeclaration, AnimationEngine() call, single-quoted
strings, MetalView unbalanced brace (F2 raycast preserved), @AppStorage
Float->Double, SceneManager instance default param, KeyframeEntry qualified
as AnimationEngine.KeyframeEntry, Scene disambiguated as SwiftUI.Scene
(Satin exports its own Scene).

Features (6+2): GestureHandlerView duplicate init overrides removed,
BooleanEngine: import OCCTSwift + qualified Shape + box() (not createBox)
+ optional unwrapping on union/subtract/intersect, PencilForceView
fileprivate, CADTimelineView drops @ObservedObject on non-observable
CADNode, AppMode referenced as top-level enum.

Sources (4+7): AnimationEngine missing closing brace, Deformer protocol
default arg moved to conformers (7 deformers updated), MaterialData
argument order, ModelLayer Codable removed (Mesh holds MTLBuffer; nothing
encodes it).

+++ ios-app/AppForgeStudio/Core/UI/PreferencesView.swift
@@ -5,7 +5,7 @@ private let logger = Logger(subsystem: "com.appforgestudio", category: "Preferen
 struct PreferencesView: View {
     @AppStorage("renderQuality") private var renderQuality: RenderQuality = .high
     @AppStorage("showGrid") private var showGrid = true
-    @AppStorage("gridSize") private var gridSize: Float = 1.0
+    @AppStorage("gridSize") private var gridSize: Double = 1.0
     @AppStorage("autoSave") private var autoSave = true
     @AppStorage("autoSaveInterval") private var autoSaveInterval: Double = 300
     @AppStorage("defaultExportFormat") private var defaultExportFormat: PreferencesExportFormat = .stl
